[SubtasksJS-6] Вынесла создание формы в отдельную функцию.
